+++ internal/environments/outputter.go
@@ -1,38 +1,31 @@
 package environments
 
-import (
-	"encoding/json"
-	"fmt"
-
-	"ldcli/internal/output"
-)
-
 // resource is the subset of data we need to display a command's plain text response.
-type resource struct {
-	Key  string `json:"key"`
-	Name string `json:"name"`
-}
+// type resource struct {
+// 	Key  string `json:"key"`
+// 	Name string `json:"name"`
+// }
 
-type EnvironmentOutputter struct {
-	resourceJSON []byte
-}
+// type EnvironmentOutputter struct {
+// 	resourceJSON []byte
+// }
 
-func (o EnvironmentOutputter) JSON() string {
-	return string(o.resourceJSON)
-}
+// func (o EnvironmentOutputter) JSON() string {
+// 	return string(o.resourceJSON)
+// }
 
-func (o EnvironmentOutputter) String() string {
-	var r resource
-	_ = json.Unmarshal(o.resourceJSON, &r)
-	fnPlaintext := func(p resource) string {
-		return fmt.Sprintf("%s (%s)", p.Name, p.Key)
-	}
+// func (o EnvironmentOutputter) String() string {
+// 	var r resource
+// 	_ = json.Unmarshal(o.resourceJSON, &r)
+// 	fnPlaintext := func(p resource) string {
+// 		return fmt.Sprintf("%s (%s)", p.Name, p.Key)
+// 	}
 
-	return output.FormatColl([]resource{r}, fnPlaintext)
-}
+// 	return output.FormatColl([]resource{r}, fnPlaintext)
+// }
 
-func NewEnvironmentOutputter(resourceJSON []byte) EnvironmentOutputter {
-	return EnvironmentOutputter{
-		resourceJSON: resourceJSON,
-	}
-}
+// func NewEnvironmentOutputter(resourceJSON []byte) EnvironmentOutputter {
+// 	return EnvironmentOutputter{
+// 		resourceJSON: resourceJSON,
+// 	}
+// }
